Fix parent environment going out of scope unexpectedly.
Unfortunately, we can't just make parent a weak pointer, as environments need to keep
their parents alive, otherwise weird stuff happens, e.g. the following fails, because
the call-site environment went out of scope, and the returned lambda didn't hold on
because of the let in-between.

(define (foo arg)
    (let ()
      (lambda () arg)))
((foo 10))